server/requirements.txt
flask-dance
openai
flask_cors
flask_login
flask_sqlalchemy
python-dotenv